+++ packages/terra-form-radio/src/RadioField.jsx
@@ -10,6 +10,7 @@ import VisualyHiddenText from 'terra-visually-hidden-text';
 import {
   VALUE_UP, VALUE_DOWN, VALUE_RIGHT, VALUE_LEFT,
 } from 'keycode-js';
+import { findFirstFocusableItem, findLastFocusableItem } from './_RadioUtil';
 import styles from './RadioField.module.scss';
 
 const cx = classNamesBind.bind(styles);
@@ -113,7 +114,7 @@ const RadioField = (props) => {
     legendAttrs.className,
   ]);
 
-  const fieldSetId = `terra-radio-group-${uniqueid()}`;
+  const fieldSetId = customProps.id || `terra-radio-group-${uniqueid()}`;
   const legendAriaDescriptionId = `terra-radio-field-description-${uniqueid()}`;
   const helpAriaDescriptionId = help ? `terra-radio-field-description-help-${uniqueid()}` : '';
   const errorAriaDescriptionId = error ? `terra-radio-field-description-error-${uniqueid()}` : '';
@@ -144,34 +145,51 @@ const RadioField = (props) => {
     </LegendGroup>
   );
 
-  const handleKeyDown = (event) => {
+  /*
+   * Note: Cyclic Navigation of Radio button is not supported in Safari browser hence adding keydown event handler to support cyclic navigation.
+   * this handler will use native mouse event to set focus back to first radio button when we press down or right arrow key on last radio button and vise versa.
+   */
+  const handleKeyDown = (event, radio) => {
     const radioGroup = document.getElementById(fieldSetId);
-    const radioItems = radioGroup.querySelectorAll('[type=radio]');
-    const itemIndex = Array.from(radioItems).indexOf(event.currentTarget);
-    if (event.key === VALUE_DOWN || event.key === VALUE_RIGHT) {
-      if (itemIndex === radioItems.length - 1) {
-        radioItems[0].focus();
-        radioItems[0].checked = true;
-      } else {
-        radioItems[itemIndex + 1].focus();
-        radioItems[itemIndex + 1].checked = true;
+    if (radioGroup) {
+      const radioItems = radioGroup.querySelectorAll('[type=radio]');
+      const itemIndex = Array.from(radioItems).indexOf(event.currentTarget);
+      const onClick = new MouseEvent('click', { bubbles: true, cancelable: false });
+      const firstItemIndex = findFirstFocusableItem(radioItems);
+      const lastItemIndex = findLastFocusableItem(radioItems);
+
+      if (event.nativeEvent.key === VALUE_DOWN || event.nativeEvent.key === VALUE_RIGHT) {
+        if (itemIndex === lastItemIndex) {
+          radioItems[firstItemIndex].dispatchEvent(onClick);
+        }
+      } else if (event.nativeEvent.key === VALUE_UP || event.nativeEvent.key === VALUE_LEFT) {
+        if (itemIndex === firstItemIndex) {
+          radioItems[lastItemIndex].dispatchEvent(onClick);
+        }
       }
-    } else if (event.key === VALUE_UP || event.key === VALUE_LEFT) {
-      if (itemIndex === 0) {
-        radioItems[radioItems.length - 1].focus();
-        radioItems[radioItems.length - 1].checked = true;
-      } else {
-        radioItems[itemIndex - 1].focus();
-        radioItems[itemIndex - 1].checked = true;
-      }
+    }
+    if (radio && radio.props.onKeyDown) {
+      radio.props.onKeyDown();
+    }
+  };
+
+  /*
+   * Focus gets lost when radio button's are selected via mouse in Safari browser.
+   * This set focus back on the radio button on mouse click
+   */
+  const handleClick = (event, radio) => {
+    event?.currentTarget?.focus();
+    if (radio && radio.props.onClick) {
+      radio.props.onClick();
     }
   };
 
   const content = React.Children.map(children, (child) => {
     if (child && child.type.isRadio) {
+      const eventHandlersForSafari = (isSafari) ? { onKeyDown: (event) => handleKeyDown(event, child), onClick: (event) => handleClick(event, child) } : undefined;
       return React.cloneElement(child, {
         inputAttrs: {
-          ...child.props.inputAttrs, 'aria-describedby': ariaDescriptionIds, onKeyDown: handleKeyDown,
+          ...child.props.inputAttrs, 'aria-describedby': ariaDescriptionIds, ...eventHandlersForSafari,
         },
       });
     }